# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/users':
    get:
      tags:
        - users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: >-
        The available information (passwords excluded) about all the
        users is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              example: [{
                id: 1,
                name: User,
                email: user.dsp@polito.it,
                self: /api/users/1
              }]
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '404':
          description: Not Found
          content: {}
  '/api/users/{userId}':
    get:
      tags:
        - 'users/{userId}'
      security:
        - cookieAuth: []
      summary: Get information about a user
      description: >-
        The available information (password excluded) about the user
        specified by userId is retrieved.
      operationId: getSingleUser
      parameters:
        - name: userId
          in: path
          description: ID of the user to get
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 1
                name: User
                email: user.dsp@polito.it
                self: /api/users/1
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '404':
          description: Not Found
          content: {}
  '/api/users/authenticator':
    post:
      tags:
        - users/authenticator
      summary: Logs a user in or out
      description: >-
        The user who wants to log in or out sends the user data to the
        authenticator which performs the operation. For logout, the operation is possible only for authenticated users.
      operationId: authenticateUser
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: The operation type ("login" or "logout")
          example: login
      requestBody:
        description: >-
          The data of the user who wants to perform log in. For login
          the structure must contain email and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              email: user.dsp@polito.it,
              password: password
        required: true
      responses:
        '204':
          description: No Content
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        '400':
          description: Bad Request
          content: 
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/JsonSchemaValidationError'
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
components:
  schemas:
    User: 
      description: schema for User data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the user
        id:
          type: integer
          description: unique identifier for the user
        email:
          type: string
          format: email
          description: user email
        name:
          type: string
          description: username
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: user password
        self: 
          type: string
          format: uri
      required:
      - email
...
    Error:
      description: schema for Error data structures
      type: object
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: object
            properties:
              param:
                type: string
                enum: ['Server']
              msg:
                description: 'Explanation of the error'
                type: string
            additionalProperties: false
            required:
              - param
              - msg
      additionalProperties: false
      required:
      - errors
    JsonSchemaValidationError:
      description: schema for JsonSchemaValidationError data structures
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          enum: ['JsonSchemaValidationError']
        validationErrors:
          type: object
          properties:
            body:
              type: array
              items:
                type: object
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid